requirements.txt
altgraph==0.17.2
aspose-words==22.5.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
future==0.18.2
google-trans-new==1.1.9
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==1.22.3
pefile==2021.9.3
Pillow==9.1.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.5
pywin32-ctypes==0.2.0
requests==2.27.1
rfc3986==1.5.0
Send2Trash==1.8.0
sniffio==1.2.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
